Refactor version check and add karma backup system
Changed the version check to use the .toc file (which only makes sense) and added a couple of functions that allow for creation and restoration of backups. It's currently in a working state where it saves at most one backup a day, but I need to figure out how to limit the number of backups. Otherwise, the back variable will go on forever.